+++ views/partials/navbar.handlebars
@@ -1,7 +1,7 @@
 <!--Main navigation bar-->
 <div id="main-nav-bar">
     <button class="tablink" onclick="location.href='/';">Home</button> <!--TODO: define action onclick https://www.w3schools.com/howto/howto_js_full_page_tabs.asp-->
-    <button class="tablink" onclick="location.href='/profile/';">Profile</button>
+    <button class="tablink" onclick="location.href='/users/';">Profile</button>
     <button class="tablink" onclick="location.href='/posts/';">Feed</button>
     <label for="nav-workspace-notification">
         <button class="tablink" onclick="location.href='/workspaces/';">Workspaces</button>
